requirements.txt
asgiref==3.7.2
boto3==1.26.155
botocore==1.29.155
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.0.0
Django==5.0.3
django-allauth==0.61.1
django-bootstrap5==23.4
django-environ==0.11.2
django-storages==1.13.2
gunicorn==20.1.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
whitenoise==6.6.0